# encoding/hex: implement hex.Decode() using LUT instead of branching