Implement hex decode using a 256 byte lookup table instead of branching logic.

In happy flow, uses 3x 64 byte (or 5x 32 byte) cache lines.
Before 
```
goos: linux
goarch: amd64
pkg: std/encoding/hex
cpu: AMD Ryzen Threadripper 3970X 32-Core Processor 
BenchmarkDecode
BenchmarkDecode/256
BenchmarkDecode/256-64         	 5013205	       226.6 ns/op	1129.81 MB/s
BenchmarkDecode/1024
BenchmarkDecode/1024-64        	 1389394	       858.3 ns/op	1193.03 MB/s
BenchmarkDecode/4096
BenchmarkDecode/4096-64        	  337632	      3431 ns/op	1193.75 MB/s
BenchmarkDecode/16384
BenchmarkDecode/16384-64       	   86403	     13837 ns/op	1184.09 MB/s
```
after:
```
goos: linux
goarch: amd64
pkg: std/encoding/hex
cpu: AMD Ryzen Threadripper 3970X 32-Core Processor 
BenchmarkDecode
BenchmarkDecode/256
BenchmarkDecode/256-64         	 8688688	       131.6 ns/op	1945.54 MB/s
BenchmarkDecode/1024
BenchmarkDecode/1024-64        	 2456132	       503.4 ns/op	2034.10 MB/s
BenchmarkDecode/4096
BenchmarkDecode/4096-64        	  594943	      1935 ns/op	2116.68 MB/s
BenchmarkDecode/16384
BenchmarkDecode/16384-64       	  152079	      7962 ns/op	2057.77 MB/s
```

Also reduces amd64 object size by 766 bytes, despite the extra RODATA due to removal of `fromHexChar()` and duplicated inlined versions of it and simplification of `Decode()`.